Callback support for atomics.
OK, I am *pretty sure* this works.. I know that the ompt_wait_id_t
is getting set, and the code is in place for the locks. However,
I can't come up with a complex enough example which requires locks.
The code provides lots of opportunity to avoid locks, so
functions like "__sync_fetch_and_add" are used instead of locked
access. for more details, see kmp_os.h

Anyway, the code for setting the wait_id should be reusable.