Add OMPT_BLAME and OMPT_TRACE flags in some files.
If a part of the code is designed specifically for implementaing
blameshift ompt feature, it's wrapped with
 #ifdef OMPT_SUPPORT && OMPT_BLAME

If the code is for ompt synchronous (which is also optional), it's
wrapped with:

 #ifdef OMPT_SUPPORT && OMPT_TRACE

Otherwise, the code is a mandatory ompt impelementation, and needs to be
wrapped with:

 #ifdef OMPT_SUPPORT

I found some codes are ambiguous, especially variable declarations in
kmp.h. Temporary, I keep them in mandatory OMPT feature until we are
sure if it's only for ompt optional implementation.