Fix bug for ompt_get_task_id within degenerate parallel region
The bug occurs when running against degenerate-parallel-region.cpp in
ompt-test-suite repository.
I think the main cause of this problem is because when lightweight task
is null, we assumed ompt_task() will retrieve the task from the
"current" task team in the "current" depth:

154   ompt_lw_taskteam_t *lwt = __ompt_get_lw_taskteam(&level);
...
163     kmp_taskdata_t *task = ompt_task(level);

However, ompt_task() doesn't keep track the current, and the right way
is to look for the task from depth 0 to the specifiec depth:

163     kmp_taskdata_t *task = ompt_task(depth);

So far this fixes when comparing the parent task id with serial task id
(inside a parallel region) and comparing the parent task id of an inner
nested region against the task id of outer parallel region.